+++ PRTelegramBot/CHANGELOG.md
@@ -3,6 +3,7 @@
 - feature: В интерфейс IInternalCheck добавлен аргумент CommandHandler
 - feature: Добавлены новые события в update типа сообщения. OnPreReplyCommandHandle, OnPostReplyCommandHandle, OnPreDynamicReplyCommandHandle, OnPostDynamicReplyCommandHandle,
     OnPreSlashCommandHandle, OnPostSlashCommandHandle, OnPreInlineCommandHandle, OnPostInlineCommandHandle, OnPreNextStepCommandHandle, OnPostNextStepCommandHandle
+- feature: UpdateExtension добавлены методы IsUserChatId, TryGetChatId
 
 -- 2024.06.30 - V0.6.1
 - update: Ядро telegram.bot обновлено с 19 до 21.2.0 версии.